Refactor: Moving utils from ibm to general
