+++ pkg/ibm/security_group.go
@@ -419,7 +419,7 @@ func IsRemoteInCIDR(remote, cidr string) (bool, error) {
 		}
 		return netCidr.Contains(netIP), nil
 	}
-	return isCIDRSubset(remote, cidr)
+	return utils.IsCIDRSubset(remote, cidr)
 }
 
 // GetRemoteType returns IBM specific keyword returned by vpc1 SDK,
