+++ pkg/utils/utils.go
@@ -14,11 +14,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package log
+package utils
 
 import (
 	"fmt"
 	"log"
+	"net"
 	"net/netip"
 	"os"
 	"strings"
@@ -175,7 +176,7 @@ func GetNumVpnConnections(cloud1, cloud2 string) int {
 
 // DoCIDROverlap returns false if cidr blocks don't share a single ip,
 // i.e. they don't overlap.
-func DoesCIDROverlap(cidr1, cidr2 string) (bool, error) {
+func DoCIDROverlap(cidr1, cidr2 string) (bool, error) {
 	netCIDR1, err := netip.ParsePrefix(cidr1)
 	if err != nil {
 		return true, err
@@ -189,4 +190,26 @@ func DoesCIDROverlap(cidr1, cidr2 string) (bool, error) {
 	}
 
 	return false, nil
+}
+
+// isCIDRSubset returns true if cidr1 is a subset (including equal) to cidr2
+func IsCIDRSubset(cidr1, cidr2 string) (bool, error) {
+	firstIP1, netCidr1, err := net.ParseCIDR(cidr1)
+	// ParseCIDR() example from Docs: for CIDR="192.0.2.1/24"
+	// IP=192.0.2.1 and network mask 192.0.2.0/24 are returned
+	if err != nil {
+		return false, err
+	}
+
+	_, netCidr2, err := net.ParseCIDR(cidr2)
+	if err != nil {
+		return false, err
+	}
+	// number of significant bits in the subnet mask
+	maskSize1, _ := netCidr1.Mask.Size()
+	maskSize2, _ := netCidr2.Mask.Size()
+	// cidr1 is a subset of cidr2 if the first user ip of cidr1 within cidr2
+	// and the network mask of cidr1 is no smaller than that of cidr2, as
+	// fewer bits are left for user address space.
+	return netCidr2.Contains(firstIP1) && maskSize1 >= maskSize2, nil
 }
